fix(global): :bug: statusbar not rendering in mobile related issue resolved
statusbar not rendering in mobile related issue resolved

ref: #45
